float-to-int

+++ libpysal/graph/base.py
@@ -390,7 +390,7 @@ class Graph(SetOpsMixin):
         from scipy import sparse
 
         if dense.dtype == bool:
-            dense = dense.astype(float)
+            dense = dense.astype(int)
 
         return cls.from_sparse(sparse.csr_array(dense), ids=ids)
